handle matching it.only/describe.only
