+++ packages/jest-editor-support/src/ScriptParser.js
@@ -25,7 +25,7 @@ class ItBlock extends Node {
   name: string;
 }
 
-type ParserReturn = { itBlocks: any,  expects: any }
+export type ParserReturn = { itBlocks: Array<ItBlock>,  expects: Array<Expect> }
 
 /**
  * Converts the file into an AST, then passes out a
