+++ packages/jest-editor-support/src/__tests__/fixtures/metaphysics/partner_show.example
@@ -1,6 +1,6 @@
 /* eslint-disable */
 
-// import moment from 'moment';
+import moment from 'moment';
 
 describe('PartnerShow type', () => {
   const PartnerShow = schema.__get__('PartnerShow');
